Update raycaster, scuffed but semi working
Rays are being casted by first checking for horizontal intersections and
then vertical intersections. The distances between the two are compared
and we used the shorter one to draw a slice of the wall. However, there
are cases where both intersections cannot be found and we have no
distances to work with, probably because it isn't proper DDA and is not
precise enough. Fix later.