Raycaster working with fisheye effect
The major problem being fixed is that the raycaster was not casting at
all. ray->hit is a boolean set for DDA to keep running as long as the
ray hasn't hit a wall yet, it was not reset after each frame, so it only
casted the first time and never again. Fixing this seemed to fix all my
confusions.