Fix textures distorting when wall height is max
Each column line was drawn from a starting position downwards to an ending
position. Because I didn't want to draw outside of the screen, I
mistakenly clamped the start and end positions into the bounds of the
screen, however this caused a weird distortion where the top of a
texture always gets drawn, even if they are supposedly "out of the
screen".