feat(config): parse map configurables into a Config
Read from the .cub file using different get_next_line "strategies".

Initially, we can skip leading and trailing whitespaces as well as
empty lines while looking for configurables such as "NO", "SO", "EA" etc
Use a get_next_line wrapper that trims away whitespaces and discards
empty lines, placing the type identifier at the beginning of a ready
line, allowing us to identify it and store it in a t_configurable.
It has ownership of the line and we store offsets to the actual value
after the type identifier.

Once done with all necessary configurables, while rejecting duplicates
if any is found, we want to skip all empty lines until we reach the map
contents. Use a get_next_line that skips all empty lines until a line
with a non-whitespace character is found. This marks the start of the
map content.

While reading the map, whitespaces and empty lines cannot be discarded
as it must be parsed as it looks like and could indicate an invalid map
configuration, so we switch to a get_next_line that only trims away the
trailing newline character and nothing else.