fix: moving across walls diagonally
If two movement keys were pressed, for example moving forward and moving
sideways, the new position was incorrect as the change in position was
accumulated for both movements and only updated afterwards, allowing one
to clip into a wall and move diagonally through it. Now, the position is
updated as soon as the new position is tested to be valid.